+++ R/gridpattern-package.R
@@ -3,6 +3,7 @@
 #'  \describe{
 #'    \item{ggpattern_array_funcs}{Set custom \dQuote{array} pattern functions.}
 #'    \item{ggpattern_geometry_funcs}{Set custom \dQuote{geometry} pattern functions.}
+#'    \item{ggpattern_res}{Set custom raster image resolution (pixels per inch) for certain patterns.}
 #'    \item{ggpattern_use_R4.1_clipping}{If `TRUE` use the grid clipping path feature introduced in R v4.1.0.
 #'                          If `FALSE` do a `rasterGrob` approximation of the clipped pattern.
 #'                          If `NULL` try to guess an appropriate choice.}