+++ R/pattern-array-plasma.R
@@ -15,7 +15,8 @@
 #' @export
 grid.pattern_plasma <- function(x = c(0, 0, 1, 1), y = c(1, 0, 0, 1), id = 1L, ...,
                                 fill = gp$fill %||% "grey80", scale = 1, alpha = gp$alpha %||% NA_real_,
-                                aspect_ratio = 1, key_scale_factor = 1, res = 72,
+                                aspect_ratio = 1, key_scale_factor = 1,
+                                res = getOption("ggpattern_res", 72),
                                 default.units = "npc", name = NULL, gp = gpar(), draw = TRUE, vp = NULL) {
     grid.pattern("plasma", x, y, id,
                  fill = fill, scale = scale, alpha = alpha,
@@ -66,7 +67,7 @@ create_magick_plasma_img <- function(width=100, height=100, colour) {
 
   #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   # Make the white transparent
-  # Colourie the black pixels into the desired colour
+  # Colorize the black pixels into the desired colour
   #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   img <- magick::image_transparent(img, 'white')
   img <- magick::image_colorize(img, opacity = 50, colour)